requirements.txt
streamlit
chromadb
dspy
mlflow
uuid
azure-identity
python-dotenv
ollama